Move github service and tests to new folder
