+++ src/DetailsView/details-view-container.tsx
@@ -20,7 +20,7 @@ import { UserConfigurationStoreData } from '../common/types/store-data/user-conf
 import { VisualizationScanResultData } from '../common/types/store-data/visualization-scan-result-data';
 import { VisualizationStoreData } from '../common/types/store-data/visualization-store-data';
 import { VisualizationType } from '../common/types/visualization-type';
-import { GitHubIssueFilingSettings } from '../issue-filing/github/github-issue-filing-service';
+import { GitHubIssueFilingSettings } from '../issue-filing/services/github/github-issue-filing-service';
 import { DetailsViewCommandBarDeps } from './components/details-view-command-bar';
 import { DetailsViewOverlay, DetailsViewOverlayDeps } from './components/details-view-overlay';
 import { DetailsRightPanelConfiguration, GetDetailsRightPanelConfiguration } from './components/details-view-right-panel';
