+++ src/injected/details-dialog-handler.ts
@@ -4,7 +4,7 @@ import { autobind } from '@uifabric/utilities';
 
 import { FeatureFlags } from '../common/feature-flags';
 import { HTMLElementUtils } from '../common/html-element-utils';
-import { GitHubIssueFilingSettings } from '../issue-filing/github/github-issue-filing-service';
+import { GitHubIssueFilingSettings } from '../issue-filing/services/github/github-issue-filing-service';
 import { UserConfigurationStoreData } from './../common/types/store-data/user-configuration-store';
 import { DetailsDialog } from './components/details-dialog';
