+++ src/issue-filing/services/github/create-github-issue-filing-url.ts
@@ -1,12 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { createIssueDetailsBuilder } from '../common/create-issue-details-builder';
-import { HTTPQueryBuilder } from '../common/http-query-builder';
-import { IssueDetailsBuilder } from '../common/issue-details-builder';
-import { MarkdownFormatter } from '../common/markup/markdown-formatter';
-import { EnvironmentInfo } from './../../common/environment-info-provider';
-import { CreateIssueDetailsTextData } from './../../common/types/create-issue-details-text-data';
-import { IssueFilingUrlStringUtils, IssueUrlCreationUtils } from './../common/issue-filing-url-string-utils';
+import { EnvironmentInfo } from '../../../common/environment-info-provider';
+import { CreateIssueDetailsTextData } from '../../../common/types/create-issue-details-text-data';
+import { createIssueDetailsBuilder } from '../../common/create-issue-details-builder';
+import { HTTPQueryBuilder } from '../../common/http-query-builder';
+import { IssueDetailsBuilder } from '../../common/issue-details-builder';
+import { IssueFilingUrlStringUtils, IssueUrlCreationUtils } from '../../common/issue-filing-url-string-utils';
+import { MarkdownFormatter } from '../../common/markup/markdown-formatter';
 import { GitHubIssueFilingSettings } from './github-issue-filing-service';
 import { rectify, UrlRectifier } from './github-url-rectifier';
