+++ src/issue-filing/services/github/github-issue-filing-service.tsx
@@ -3,10 +3,11 @@
 import { isEmpty, isString } from 'lodash';
 import { TextField } from 'office-ui-fabric-react/lib/TextField';
 import * as React from 'react';
-import { NamedSFC } from '../../common/react/named-sfc';
-import { createSettingsGetter } from '../common/create-settings-getter';
-import { IssueFilingService } from '../types/issue-filing-service';
-import { SettingsFormProps } from '../types/settings-form-props';
+
+import { NamedSFC } from '../../../common/react/named-sfc';
+import { createSettingsGetter } from '../../common/create-settings-getter';
+import { IssueFilingService } from '../../types/issue-filing-service';
+import { SettingsFormProps } from '../../types/settings-form-props';
 import { gitHubIssueFilingUrlProvider } from './create-github-issue-filing-url';
 
 const GitHubIssueFilingServiceKey = 'gitHub';
