+++ src/issue-filing/services/null-issue-filing-service/null-issue-filing-service.tsx
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { NamedSFC } from '../../../common/react/named-sfc';
-import { SettingsFormProps } from '../../types/settings-form-props';
 import { IssueFilingService } from '../../types/issue-filing-service';
+import { SettingsFormProps } from '../../types/settings-form-props';
 
 const nullServiceKey = 'none';
