+++ src/tests/unit/tests/DetailsView/details-view-container.test.tsx
@@ -37,7 +37,7 @@ import {
 import { DetailsViewMainContent } from '../../../../DetailsView/details-view-main-content';
 import { DetailsViewToggleClickHandlerFactory } from '../../../../DetailsView/handlers/details-view-toggle-click-handler-factory';
 import { PreviewFeatureFlagsHandler } from '../../../../DetailsView/handlers/preview-feature-flags-handler';
-import { GitHubIssueFilingSettings } from '../../../../issue-filing/github/github-issue-filing-service';
+import { GitHubIssueFilingSettings } from '../../../../issue-filing/services/github/github-issue-filing-service';
 import { DetailsViewStoreDataBuilder } from '../../common/details-view-store-data-builder';
 import { TabStoreDataBuilder } from '../../common/tab-store-data-builder';
 import { CreateTestAssessmentProviderWithFeatureFlag } from '../../common/test-assessment-provider';
