+++ src/tests/unit/tests/issue-filing/services/github/create-github-issue-filing-url.test.ts
@@ -2,14 +2,14 @@
 // Licensed under the MIT License.
 import { IMock, Mock } from 'typemoq';
 
-import { HTTPQueryBuilder } from '../../../../../issue-filing/common/http-query-builder';
-import { IssueDetailsBuilder } from '../../../../../issue-filing/common/issue-details-builder';
-import { createGitHubIssueFilingUrlProvider } from '../../../../../issue-filing/github/create-github-issue-filing-url';
-import { UrlRectifier } from '../../../../../issue-filing/github/github-url-rectifier';
-import { IssueFilingUrlProvider } from '../../../../../issue-filing/types/issue-filing-service';
-import { EnvironmentInfo } from './../../../../../common/environment-info-provider';
-import { IssueUrlCreationUtils } from './../../../../../issue-filing/common/issue-filing-url-string-utils';
-import { GitHubIssueFilingSettings } from './../../../../../issue-filing/github/github-issue-filing-service';
+import { EnvironmentInfo } from '../../../../../../common/environment-info-provider';
+import { HTTPQueryBuilder } from '../../../../../../issue-filing/common/http-query-builder';
+import { IssueDetailsBuilder } from '../../../../../../issue-filing/common/issue-details-builder';
+import { IssueUrlCreationUtils } from '../../../../../../issue-filing/common/issue-filing-url-string-utils';
+import { createGitHubIssueFilingUrlProvider } from '../../../../../../issue-filing/services/github/create-github-issue-filing-url';
+import { GitHubIssueFilingSettings } from '../../../../../../issue-filing/services/github/github-issue-filing-service';
+import { UrlRectifier } from '../../../../../../issue-filing/services/github/github-url-rectifier';
+import { IssueFilingUrlProvider } from '../../../../../../issue-filing/types/issue-filing-service';
 
 const buildedUrl = 'https://builded-url';
 describe('createGitHubIssueFilingUrlTest', () => {
